Added type tags to hats
The Plan for the future is to only use CustomModelData IDs in the loot tables defining the hats.
Instead unique tags of the format "hats.hat.type.xxxx" are going to take their place.
So selectors like @e[nbt={... CustomModelData=xxxx ...}] will converto to @e[nbt={... Tags:["hats.hat.type.xxxx"] ...}].
That would make the datapack much easier to maintain.

Since the changing of selectors will make existing hats unusable, this change prepares the hats with the type tags, so that in a later update the selectors can be changed with less problems for users.