Hat fixing now ignores hats on the Players head
Previously, if the Player put on a hat and it would be converted to a stick hat, the fixing process was triggered. Thats not necessarily bad, because it stops after checking for stick hats in the hotbar and inventory, but it still executes all 34 inventory copying commands.

Instead, the process now immediately fails if there is only a stick hat in the Players head slot.